harvester/requirements.txt
# Harvester web
django==4.2.15
djangorestframework==3.15.2
django-cors-headers==3.14.0
whitenoise==6.7.0
psycopg2-binary==2.9.9
Pillow==10.4.0
python-dateutil==2.9.0.post0
uritemplate==3.0.1
inflection==0.5.1
Jinja2==3.1.4
django-x-robots-tag-middleware==1.3.3
markupsafe==2.1.5
django-debug-toolbar==4.4.6
pymsteams==0.2.3
django-admin-confirm==1.0.0
django-log-formatter-json==0.2.0
ujson==3.2.0  # implicit dependency of django-log-formatter-json

# Harvester Celery
celery==5.4.0
celery[redis]
redis==4.6.0
flower==2.0.1

# Harvester search
opensearch-py==2.7.1

# Harvester other
vobject==0.9.7
pdf2image==1.17.0
Unidecode==1.3.8
django-versatileimagefield==3.1
django-mptt==0.16.0

# Datagrowth & dependencies
datagrowth==0.19.8
requests==2.32.3
urllib3==1.26.19
jsonschema==3.2.0
pytz==2021.3
urlobject==2.4.3
lxml==4.9.4
pandas==2.2.2
beautifulsoup4==4.12.3

# AWS
awscli==1.34.10
boto3==1.35.10
requests-aws4auth==1.3.1
django-storages==1.14.4

# Tools
watchdog==2.3.1
ipython==8.26.0
tqdm==4.66.5
invoke==2.2.0
ipdb==0.13.13
sentry-sdk==1.45.1

# Patched packages
youtube-dl@git+https://github.com/ytdl-org/youtube-dl.git@master
git+https://github.com/fako/django-json-field.git
